refactor: extract RecRequestFactory from RecMaster in distributed runtime.
Move Rec request-building logic out of RecMaster into a dedicated
RecRequestFactory. The factory owns the pipeline strategy hierarchy
(LlmRec, LlmRec-with-mm-data, OneRec prefill-only, OneRec xattention),
the input-processing helpers, and build_request_common. RecMaster now
owns the factory and delegates from its handle_request overloads while
retaining request scheduling.
