+++ xllm/core/framework/request/CMakeLists.txt
@@ -12,6 +12,7 @@ cc_library(
     request.h
     llm_request_factory.h
     vlm_request_factory.h
+    rec_request_factory.h
     dit_request.h
     request_output.h
     usage.h
@@ -35,6 +36,7 @@ cc_library(
     request_base.cpp
     llm_request_factory.cpp
     vlm_request_factory.cpp
+    rec_request_factory.cpp
     dit_request.cpp
     request_output.cpp
     dit_request_output.cpp
